requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
gunicorn>=21.2
whitenoise>=6.5
